Fix bug where the player would bet after being dealt the cards (#8)
There was a bug where the player would bet after being dealt the cards. The player wouldn't use the cards to adjust their bet according to the specific EV of this combination, but rather it affected the calculation of the true count which affected the bet.

It also includes some minor fixes for splitting, and improves the calculation of the risk of ruin.